xr.infer_freq (#4033)
* xr.infer_freq and related code

* Formatting and comments

* Rewrite _CFTimeFrequencyInferer independently of pandas

* Syntax and add frequency.py file

* Fix tests and month_deltas

* Require cftime 1.1.0 for the test

* Apply suggestions from code review


* Changes following code review

* Docs

* Docs

* Black

* Fix tests for requiring cftime 1.1.0

* Update whats-new

* Apply suggestions from code review


* Add invalid input tests for better coverage

* Fix link in whats-new.rst

+++ xarray/coding/cftimeindex.py
@@ -578,7 +578,8 @@ class CFTimeIndex(pd.Index):
             [
                 _total_microseconds(exact_cftime_datetime_difference(epoch, date))
                 for date in self.values
-            ]
+            ],
+            dtype=np.int64,
         )
 
     def _round_via_method(self, freq, method):
